requirements.txt
# Agente de E-Commerce — Dependências
# Instale com: pip install -r requirements.txt

# ── Core do Agente ────────────────────────────────────────
google-generativeai>=0.8.0      # SDK Gemini (Gemini 2.5 Flash)
pandas>=2.0.0                   # Manipulação de dados
tabulate>=0.9.0                 # Formatação de tabelas
python-dotenv>=1.0.0            # Carrega variáveis do .env

# ── API FastAPI ───────────────────────────────────────────
fastapi>=0.115.0                # Framework web
uvicorn[standard]>=0.30.0      # Servidor ASGI

# ── Jupyter Notebook ──────────────────────────────────────
jupyter>=1.0.0
ipykernel>=6.0.0
